divrep coul: don't subtract uints
This doesn't appear to have caused problems, but it's safer to cast
these small uints to int before subtracting when we want a signed
result.

+++ divrep/coul.c
@@ -596,7 +596,7 @@ void recover(void) {
 
 int cmp_high(const void *va, const void *vb) {
     uint a = *(uint *)va, b = *(uint *)vb;
-    return divisors[b].high - divisors[a].high;
+    return (int)divisors[b].high - (int)divisors[a].high;
 }
 
 /* TODO: use MPU code for ulong next_prime */
